Reorganize Project Structure: Move Scripts to Pipelines Directory
- Migrate scripts from `scripts/` to `pipelines/` directory
- Update import paths in `main.py` to reflect new directory structure
- Remove deprecated analysis and visualization scripts
- Simplify project layout by consolidating related scripts
