+++ scripts/inspect/analyze_faiss_data.py
