+++ scripts/inspect/analyze_other_agencies.py
