+++ scripts/inspect/analyze_word_counts.py
