+++ scripts/inspect/check_db.py
