+++ scripts/inspect/validate_chunking.py
