+++ scripts/inspect/view_embeddings_db.py
