+++ scripts/inspect/view_regulations_db.py
